+++ lib/chequeo/healthchecks/model_validity.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Chequeo
+  module HealthChecks
+    class ModelValidity < Base
+
+    end
+  end
+end