+++ lib/chequeo/strategy/base.rb
@@ -1,7 +1,0 @@
-module Chequeo
-  module Strategy
-    class Base
-
-    end
-  end
-end