+++ lib/generators/chequeo/templates/initializer.rb
@@ -1,0 +1,3 @@
+Chequeo.configure do |config|
+  # config.logger = Rails.logger
+end