+++ test/lib/chequeo/test_helper.rb
@@ -1,0 +1,5 @@
+require 'chequeo'
+require 'minitest/autorun'
+
+require 'simplecov'
+SimpleCov.start