+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+## 0.2.0.beta
+- Beta Release of application. Expect breaking changes in the future.
+
+## 0.1.0
+- Create simple daemon
+
 ## 0.0.1
 - Initial version created